use ssize_t everywhere

+++ maps/include/maps/FlatSkyMap.h
@@ -112,25 +112,25 @@ public:
 	double yres() const;
 	double res() const;
 
-	size_t AngleToPixel(double alpha, double delta) const override;
+	ssize_t AngleToPixel(double alpha, double delta) const override;
 	std::vector<double> PixelToAngle(size_t pixel) const override;
 	std::vector<double> PixelToAngle(size_t x_pix, size_t y_pix) const;
 	std::vector<double> PixelToAngleWrapRa(size_t pixel) const;
 	std::vector<double> AngleToXY(double alpha, double delta) const;
 	std::vector<double> XYToAngle(double x, double y) const;
-	size_t XYToPixel(double x, double y) const;
+	ssize_t XYToPixel(double x, double y) const;
 	std::vector<double> PixelToXY(size_t pixel) const;
 	std::vector<double> QuatToXY(quat q) const;
 	quat XYToQuat(double x, double y) const;
 
 	std::vector<double> PixelToAngleGrad(size_t pixel, double h=0.001) const;
 
-	void GetRebinAngles(long pixel, size_t scale,
+	void GetRebinAngles(size_t pixel, size_t scale,
 	    std::vector<double> & alphas, std::vector<double> & deltas) const override;
 	void GetInterpPixelsWeights(double alpha, double delta,
-	    std::vector<long> & pixels, std::vector<double> & weights) const override;
+	    std::vector<ssize_t> & pixels, std::vector<double> & weights) const override;
 
-	std::vector<long> QueryDisc(double alpha, double delta,
+	std::vector<ssize_t> QueryDisc(double alpha, double delta,
 	    double radius) const override;
 
 	G3SkyMapPtr Rebin(size_t scale, bool norm = true) const override;
